Refactor: approvalweightmanager models
